backend/photos/serializers.py
from rest_framework import serializers
from .models import Photo

# code here.
class PhotoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Photo
        fields = '__all__'

class PhotoDetailSerializer(serializers.ModelSerializer):
    class Meta:
        model = Photo
        fields = ['id', 'image_url', 'is_used', 'created_at', 'used_at']