backend/requirements.txt
asgiref==3.9.1
boto3==1.40.8
botocore==1.40.8
certifi==2025.7.9
charset-normalizer==3.4.2   
drf_yasg
Django==5.2.4
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework-simplejwt==5.2.2
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
